move frontends/ to top level
